requirements.txt
fastapi
uvicorn[standard]
mlflow
joblib
pyyaml
dvc
